+++ unity/theme-research/Assets/Scripts/WhisperClasses.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f161b0e21405e1f409acb31a3d986708
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
